+++ src/Deedle.Math/AssemblyInfo.fs
@@ -2,18 +2,18 @@
 namespace System
 open System.Reflection
 
-[<assembly: AssemblyTitleAttribute("Deedle")>]
-[<assembly: AssemblyProductAttribute("Deedle")>]
-[<assembly: AssemblyDescriptionAttribute("Easy to use .NET library for data manipulation and scientific programming")>]
+[<assembly: AssemblyTitleAttribute("Deedle.Math")>]
+[<assembly: AssemblyProductAttribute("Deedle.Math")>]
+[<assembly: AssemblyDescriptionAttribute("Deedle interop with Math.Net")>]
 [<assembly: AssemblyVersionAttribute("2.1.0")>]
 [<assembly: AssemblyInformationalVersionAttribute("2.1.0")>]
 [<assembly: AssemblyFileVersionAttribute("2.1.0")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] AssemblyTitle = "Deedle"
-    let [<Literal>] AssemblyProduct = "Deedle"
-    let [<Literal>] AssemblyDescription = "Easy to use .NET library for data manipulation and scientific programming"
+    let [<Literal>] AssemblyTitle = "Deedle.Math"
+    let [<Literal>] AssemblyProduct = "Deedle.Math"
+    let [<Literal>] AssemblyDescription = "Deedle interop with Math.Net"
     let [<Literal>] AssemblyVersion = "2.1.0"
     let [<Literal>] AssemblyInformationalVersion = "2.1.0"
     let [<Literal>] AssemblyFileVersion = "2.1.0"