+++ src/Deedle/FSharp.Data/Csv/CsvRuntime.fs
@@ -353,7 +353,7 @@ type CsvFile<'RowType> private (rowToStringArray:Func<'RowType,string[]>, dispos
 
   /// Saves CSV to the specified file
   member x.Save(path:string, [<Optional>] ?separator, [<Optional>] ?quote) =
-    use writer = new StreamWriter(File.OpenWrite(path))
+    use writer = new StreamWriter(path)
     x.Save(writer, ?separator=separator, ?quote=quote)
 
   /// Saves CSV to a string